requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pymongo==4.6.3
fastapi==0.115.4
pydantic==2.9.2
uvicorn==0.32.0